feat: new command for editing mr descriptions
Changing a merge request description via the update command
is not very handy.  The new command "glab mr edit"
opens the merge request's description in the
editor and makes an update on save and exit.

The proposed command "glab mr 123 update --description" in
issue #466 was dismissed, see profclems response why:

> Since --description is a string flag, passing it without
> argument is an error unless the user passes an empty string
> --description ""